+++ src/clients/symbiote/index.js
@@ -1,9 +1,13 @@
 import { getOptions, saveOption } from "~/utils/storage";
 
-getOptions();
+const main = async () => {
+  await getOptions();
 
-const href = window.location.href;
-saveOption("symbioteURL", href.substring(0, href.lastIndexOf("/")));
+  const href = window.location.href;
+  await saveOption("symbioteURL", href.substring(0, href.lastIndexOf("/")));
 
-window.location.href =
-  "https://www.dndbeyond.com/sign-in?returnUrl=/characters";
+  window.location.href =
+    "https://www.dndbeyond.com/sign-in?returnUrl=/characters";
+};
+
+main();